+++ packages/components-native/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.98.2](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.98.1...@jobber/components-native@0.98.2) (2026-03-18)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
 ## [0.98.1](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.98.0...@jobber/components-native@0.98.1) (2026-03-13)
 
 **Note:** Version bump only for package @jobber/components-native