+++ packages/components/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.116.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@6.116.0...@jobber/components@6.116.1) (2026-03-18)
+
+
+### Bug Fixes
+
+* **components:** overlay SegmentedControl selected border on container border ([#2992](https://github.com/GetJobber/atlantis/issues/2992)) ([62bda36](https://github.com/GetJobber/atlantis/commit/62bda36d4a51ac44dece4b186016b4f395487335))
+
+
+
+
+
+# [6.116.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@6.115.3...@jobber/components@6.116.0) (2026-03-18)
+
+
+### Features
+
+* **components:** Add fullScreen size to Modal ([#2979](https://github.com/GetJobber/atlantis/issues/2979)) ([213bc6a](https://github.com/GetJobber/atlantis/commit/213bc6ab0aa0efcb4ac0f5f9ef6cd52a1e500a10))
+
+
+
+
+
 ## [6.115.3](https://github.com/GetJobber/atlantis/compare/@jobber/components@6.115.2...@jobber/components@6.115.3) (2026-03-16)
 
 **Note:** Version bump only for package @jobber/components